Split exonum_derive, change exonum_derive version
