+++ exonum/Cargo.toml
@@ -56,7 +56,7 @@ protobuf = "=2.2.0"
 exonum-crypto = { version = "0.9.0", path = "../crypto" }
 exonum_rocksdb = "0.7.4"
 exonum_sodiumoxide = { version = "0.0.20", optional = true }
-exonum_derive = { path = "../exonum_derive" }
+exonum_derive = { version = "0.9.0", path = "../exonum_derive" }
 
 [dev-dependencies]
 pretty_assertions = "=0.5.1"
