+++ exonum_derive/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "exonum_derive"
-version = "0.1.0"
+version = "0.9.0"
 authors = ["The Exonum Team <exonum@bitfury.com>"]
 
 [lib]
